Develop to master

+++ src/Spinner/Core/WidthDeterminer.php
@@ -1,7 +1,7 @@
 <?php
 
 declare(strict_types=1);
-// 02.03.23
+
 namespace AlecRabbit\Spinner\Core;
 
 use AlecRabbit\Spinner\Mixin\NoInstanceTrait;
@@ -11,7 +11,6 @@ use function AlecRabbit\WCWidth\wcswidth;
 use function mb_strlen;
 use function strlen;
 
-/** @internal */
 final class WidthDeterminer
 {
     use NoInstanceTrait;
@@ -23,9 +22,12 @@ final class WidthDeterminer
         if (null === self::$determiner) {
             self::$determiner = self::createDeterminer();
         }
-        return (self::$determiner)($string);
+        return (int)(self::$determiner)($string);
     }
 
+    /**
+     * @codeCoverageIgnore
+     */
     private static function createDeterminer(): Closure
     {
         if (function_exists('\AlecRabbit\WCWidth\wcswidth')) {
